requirements/base.txt
Django==4.2.4
channels==4.0.0
daphne==4.1.0
channels-redis==4.2.0
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0

celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0

django-unfold==0.45.0

whitenoise==6.6.0

django-filter==23.5
django-extensions==3.2.3
django-cors-headers==4.3.1
django-storages==1.14.2

drf-jwt==1.19.2

django-parler==2.3
django-rosetta==0.10.1

boto3==1.34.42
attrs==23.2.0

djangorestframework-simplejwt==5.3.1
google-api-python-client==2.124.0
PyJWT==2.8.0
python-social-auth==0.3.6
drf-spectacular==0.27.1

django-redis==5.4.0

pillow==10.2.0

ipython==8.14.0

django-thumbnails==0.8.0
easy-thumbnails==2.9
